NMS-12383: refactor package structure
