+++ features/timeseries/src/main/java/org/opennms/netmgt/timeseries/integration/TimeseriesFetchStrategyProvider.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.measurements.impl;
+package org.opennms.netmgt.timeseries.integration;
 
 import org.opennms.netmgt.measurements.api.MeasurementFetchStrategy;
 import org.opennms.netmgt.measurements.api.MeasurementFetchStrategyProvider;
