+++ features/timeseries/src/main/java/org/opennms/netmgt/timeseries/integration/persistence/TimeseriesPersisterFactory.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.collection.persistence.timeseries;
+package org.opennms.netmgt.timeseries.integration.persistence;
 
 import java.util.Objects;
