+++ features/timeseries/src/main/resources/META-INF/services/org.opennms.netmgt.measurements.api.MeasurementFetchStrategyProvider
@@ -1,1 +1,1 @@
-org.opennms.netmgt.measurements.impl.TimeseriesFetchStrategyProvider
+org.opennms.netmgt.timeseries.integration.TimeseriesFetchStrategyProvider
