+++ features/timeseries/src/test/java/org/opennms/netmgt/timeseries/integration/persistence/TimeseriesPersisterIT.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.collection.persistence.timeseries;
+package org.opennms.netmgt.timeseries.integration.persistence;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
